+++ host/index.jsx
@@ -168,6 +168,7 @@
   loadCmd('commands/import/text.jsx');
   loadCmd('commands/import/adjust.jsx');
   loadCmd('commands/import/fonts.jsx');
+  loadCmd('commands/import/audit.jsx');
 
   $.__rebound.loadErrors = loadErrors;
   $.__rebound.loaded = true;